bugfix: fix --from can not pull image from remote registry
When use --from to pull image from remote registry, it will reuse the systemContext from previous FROM stage. If the two image is different, the --from image will not be pulled since it use the wrong cert path. Besides, if the transport's from is "docker", the pattern "://" should be append instead of ":"

Signed-off-by: DCCooper <1866858@gmail.com>