isula-build: change default umask of isula-builder process
As the daemon process, the working umask 0027 is not suitable.
0022 is better.

Take this example:
isula-builder using 0027, the working dir of container for RUN command
(overlay/merged dir, as an example), has the 0640. Then developer changes
working user in RUN command to a non-root user, it will not have
correct permission to work.

Signed-off-by: Lu Jingxiao <lujingxiao@huawei.com>